services/wallet_service/pyproject.toml
[project]
name = "wallet-service"
version = "0.1.0"
description = "Wallet ledger service for fintech practice stack"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "sqlalchemy",
    "alembic",
    "asyncpg",
    "psycopg[binary]",
    "pydantic",
    "pydantic-settings",
    "redis",
    "python-jose[cryptography]",
    "loguru",
    "prometheus-client",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-sdk"
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
